SDIT-1490 Refactor
